Move Player extensions to Summoning.kt and timers to their own file
