+++ game/src/main/kotlin/content/skill/summoning/Summoning.kts
@@ -2,44 +2,18 @@ package content.skill.summoning
 
 import content.entity.player.dialogue.type.choice
 import content.entity.player.inv.inventoryItem
-import world.gregs.voidps.cache.definition.data.NPCDefinition
 import world.gregs.voidps.engine.client.message
 import world.gregs.voidps.engine.client.ui.interfaceOption
 import world.gregs.voidps.engine.data.definition.EnumDefinitions
-import world.gregs.voidps.engine.data.definition.ItemDefinitions
 import world.gregs.voidps.engine.data.definition.NPCDefinitions
-import world.gregs.voidps.engine.entity.character.mode.Follow
-import world.gregs.voidps.engine.entity.character.move.tele
-import world.gregs.voidps.engine.entity.character.npc.NPC
-import world.gregs.voidps.engine.entity.character.npc.NPCs
-import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.player.skill.Skill
-import world.gregs.voidps.engine.entity.item.Item
 import world.gregs.voidps.engine.entity.playerSpawn
 import world.gregs.voidps.engine.inject
 import world.gregs.voidps.engine.inv.inventory
 import world.gregs.voidps.engine.inv.remove
-import world.gregs.voidps.engine.queue.softQueue
-import world.gregs.voidps.engine.timer.timerStart
-import world.gregs.voidps.engine.timer.timerStop
-import world.gregs.voidps.engine.timer.timerTick
 
 val enums: EnumDefinitions by inject()
-val npcs: NPCs by inject()
 val npcDefinitions: NPCDefinitions by inject()
-val itemDefinitions: ItemDefinitions by inject()
-
-var Player.follower: NPC?
-    get() {
-        val index = this["follower_index", -1]
-        return world.gregs.voidps.engine.get<NPCs>().indexed(index)
-    }
-    set(value) {
-        if (value != null) {
-            this["follower_index"] = value.index
-            this["follower_id"] = value.id
-        }
-    }
 
 inventoryItem("Summon", "*_pouch") {
     val familiarLevel = enums.get("summoning_pouch_levels").getInt(item.def.id)
@@ -128,109 +102,4 @@ playerSpawn {player ->
     player.variables.send("follower_details_chathead_animation")
     player.timers.restart("familiar_timer")
     player.summonFamiliar(familiarDef, true)
-}
-
-fun Player.summonFamiliar(familiar: NPCDefinition, restart: Boolean): NPC? {
-    if (follower != null) {
-        // TODO: Find actual message for this
-        message("You must dismiss your current follower before summoning another.")
-        return null
-    }
-
-    val familiarNpc = npcs.add(familiar.stringId, tile)
-    familiarNpc.mode = Follow(familiarNpc, this)
-
-    softQueue("summon_familiar", 2) {
-        follower = familiarNpc
-
-        follower!!.gfx("summon_familiar_size_${follower!!.size}")
-        player.updateFamiliarInterface()
-        if(!restart) timers.start("familiar_timer")
-    }
-
-    return familiarNpc
-}
-
-fun Player.dismissFamiliar() {
-    npcs.remove(follower)
-    follower = null
-    interfaces.close("familiar_details")
-
-    this["follower_details_name"] = 0
-    this["follower_details_chathead"] = 0
-    this["familiar_details_minutes_remaining"] = 0
-    this["familiar_details_seconds_remaining"] = 0
-    timers.stop("familiar_timer")
-}
-
-fun Player.updateFamiliarInterface() {
-    if (follower == null) return
-
-    this.interfaces.open("familiar_details")
-
-    this["follower_details_name"] = enums.get("summoning_familiar_ids").getKey(follower!!.def.id)
-    this["follower_details_chathead"] = follower!!.def.id
-
-    this["follower_details_chathead_animation"] = 1
-}
-
-fun Player.openFollowerLeftClickOptions() {
-    interfaces.open("follower_left_click_options")
-}
-
-fun Player.confirmFollowerLeftClickOptions() {
-    this["summoning_orb_left_click_option"] = this["summoning_menu_left_click_option", -1]
-    interfaces.close("follower_left_click_options")
-}
-
-fun Player.callFollower() {
-    follower!!.tele(steps.follow, clearMode = false)
-    follower!!.clearWatch()
-}
-
-fun Player.renewFamiliar() {
-    val pouchId = enums.get("summoning_familiar_ids").getKey(follower!!.def.id)
-    val pouchItem = Item(itemDefinitions.get(pouchId).stringId)
-
-    if (!inventory.contains(pouchItem.id)) {
-        // TODO: Find the actual message used here in 2011
-        message("You don't have the required pouch to renew your familiar.")
-        return
-    }
-
-    inventory.remove(pouchItem.id)
-    this["familiar_details_minutes_remaining"] = follower!!.def["summoning_time_minutes", 0]
-    this["familiar_details_seconds_remaining"] = 0
-    follower!!.gfx("summon_familiar_size_${follower!!.size}")
-}
-
-timerStart("familiar_timer") {player ->
-    interval = 50 // 30 seconds
-
-    if(!restart) {
-        player["familiar_details_minutes_remaining"] = player.follower!!.def["summoning_time_minutes", 0]
-        player["familiar_details_seconds_remaining"] = 0
-    }
-}
-
-timerTick("familiar_timer") {player ->
-    if (player["familiar_details_seconds_remaining", 0] == 0) {
-        player.dec("familiar_details_minutes_remaining")
-    }
-    player["familiar_details_seconds_remaining"] = (player["familiar_details_seconds_remaining", 0] + 1) % 2
-
-    if (player["familiar_details_seconds_remaining", 0] <= 0 && player["familiar_details_minutes_remaining", 0] <= 0) {
-        cancel()
-    }
-}
-
-timerStop("familiar_timer") {player ->
-    if (logout) {
-        npcs.remove(player.follower)
-        return@timerStop
-    }
-
-    if (player.follower != null) {
-        player.dismissFamiliar()
-    }
 }
